requirements.txt
asgiref==3.3.4
astroid==2.5.3
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.2
django-js-asset==1.2.2
django-mptt==0.12.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
idna==2.10
isort==5.8.0
itypes==1.2.0
Jinja2==2.11.3
lazy-object-proxy==1.6.0
MarkupSafe==1.1.1
mccabe==0.6.1
PyJWT==2.0.1
pylint==2.7.4
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.4
wrapt==1.12.1
djangorestframework-recursive==0.1.2